requirements.txt
ultralytics
opencv-python
deep-sort-realtime
networkx
streamlit
numpy
pandas
pyyaml
